pyproject.toml
[project]
name = "smoke-detection-iot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.29.5",
    "paho-mqtt>=2.1.0",
    "pandas>=2.2.3",
]
